Change mapping somewhat to deal with unusable slot 51...
Need a slot for jackets as they're rather common. Made room by using slot
40 which was previously for right leg. Don't really like adapting a
vanilla slot this way but it seems the vast majority or all items using
slot 40 also use slot 39 (left leg). So merging both legs into slot 39
should "probably" be safe. This does create a minor conflict with vanilla
(can't use jacket with vanilla pants). But that is a seemingly rare
occurrence so won't worry about it; can revisit it later if it becomes a
problem.